Heroku main server/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
nemo-emblems==4.4.0
netifaces==0.10.4
numpy==1.19.5
onboard==1.4.1
opencv-python==4.5.2.54
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparted==3.11.1
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-debian==0.1.32
python-xapp==1.8.1
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
system-service==0.3
systemd-python==234
tinycss==0.4
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.0.22
urllib3==1.22
xkit==0.0.0
pytesseract==0.3.8
transformers~=4.8.2
